Update Extras.yml
Hi Santiago!
It's my first PR in my life so I hope I am doing it right.

I noticed that the Extras CF had a false positive on one of my releases of "E.T. the Extra-Terrestial". It was mainly due to poor naming of one of my spanish indexers but I thought it could be improved to avoid it, even when the name of the movie is tricky.

I basically merged a combination between your's and TRaSH's. Let me know it you interested!